# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SysCont Web - Agência Espacial Brasileira - Sistema de Gestão de Contratos Administrativos
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contratos/{id}/medicoes:
    get:
      operationId: listMedicoesByContrato
      tags: [medicoes]
      summary: Listar medições do contrato
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Lista de medições
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicao"

  # ──────────────────────────────
  # FORNECEDORES
  # ──────────────────────────────
...
  /medicoes:
    post:
      operationId: createMedicao
      tags: [medicoes]
      summary: Registrar medição
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicaoInput"
      responses:
        "201":
          description: Medição criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicao"
  /medicoes/{id}:
    get:
      operationId: getMedicao
      tags: [medicoes]
      summary: Obter medição
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Medição
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicao"
    patch:
      operationId: updateMedicao
      tags: [medicoes]
      summary: Atualizar medição
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicaoUpdate"
      responses:
        "200":
          description: Medição atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicao"
    delete:
      operationId: deleteMedicao
      tags: [medicoes]
      summary: Excluir medição
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removido
  /medicoes/{id}/aprovar:
    patch:
      operationId: aprovarMedicao
      tags: [medicoes]
      summary: Aprovar medição
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Medição aprovada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicao"
  /medicoes/{id}/pagar:
    patch:
      operationId: pagarMedicao
      tags: [medicoes]
      summary: Registrar pagamento da medição
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicaoPagamentoInput"
      responses:
        "200":
          description: Medição paga
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicao"

  # ──────────────────────────────
  # ALERTAS
  # ──────────────────────────────
...
  /pncp/buscar:
    get:
      operationId: buscarPncp
      tags: [pncp]
      summary: Buscar contratos no PNCP
      parameters:
        - name: cnpj
          in: query
          schema: { type: ["string", "null"] }
        - name: dataInicial
          in: query
          required: true
          schema: { type: string, description: "Formato YYYYMMDD" }
        - name: dataFinal
          in: query
          required: true
          schema: { type: string, description: "Formato YYYYMMDD" }
        - name: pagina
          in: query
          schema: { type: integer, default: 1 }
        - name: tamanhoPagina
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Contratos encontrados no PNCP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PncpResultado"
  /pncp/importar:
    post:
      operationId: importarPncp
      tags: [pncp]
      summary: Importar contrato do PNCP para o sistema local
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PncpImportInput"
      responses:
        "201":
          description: Contrato importado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contrato"
  /pncp/orgao/{cnpj}:
    get:
      operationId: buscarOrgaoPncp
      tags: [pncp]
      summary: Buscar informações de órgão no PNCP pelo CNPJ
      parameters:
        - name: cnpj
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Dados do órgão
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgaoPncp"

  # ──────────────────────────────
  # COMPRASNET
  # ──────────────────────────────
...
components:
  schemas:
...
    Contrato:
      type: object
      required:
        - id
        - numeroContrato
        - objeto
        - fornecedorId
        - valorInicial
        - valorAtual
        - valorPago
        - percentualExecucao
        - dataAssinatura
        - dataVigenciaInicio
        - dataVigenciaFim
        - status
        - modalidade
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        numeroContrato: { type: string }
        numeroControlePncp: { type: ["string", "null"] }
        processo: { type: ["string", "null"] }
        objeto: { type: string }
        fornecedorId: { type: integer }
        fornecedor: { $ref: "#/components/schemas/Fornecedor" }
        valorInicial: { type: number }
        valorAtual: { type: number }
        valorPago: { type: number }
        percentualExecucao: { type: number }
        fiscal: { type: ["string", "null"] }
        gestor: { type: ["string", "null"] }
        dataAssinatura: { type: string, format: date }
        dataVigenciaInicio: { type: string, format: date }
        dataVigenciaFim: { type: string, format: date }
        status:
          type: string
          enum: [vigente, encerrado, rescindido, suspenso]
        modalidade:
          type: string
          enum:
            - pregao_eletronico
            - pregao_presencial
            - concorrencia
            - tomada_de_precos
            - convite
            - dispensa
            - inexigibilidade
            - outros
        categoriaProcesso: { type: ["string", "null"] }
        unidadeGestora: { type: ["string", "null"] }
        observacoes: { type: ["string", "null"] }
        numeroParcelas: { type: ["integer", "null"] }
        aditivos:
          type: array
          items:
            $ref: "#/components/schemas/Aditivo"
        medicoes:
          type: array
          items:
            $ref: "#/components/schemas/Medicao"
        alertas:
          type: array
          items:
            $ref: "#/components/schemas/Alerta"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Fornecedor:
      type: object
      required: [id, nome, cnpjCpf, tipoPessoa, createdAt]
      properties:
        id: { type: integer }
        nome: { type: string }
        cnpjCpf: { type: string }
        tipoPessoa:
          type: string
          enum: [PJ, PF]
        email: { type: ["string", "null"] }
        telefone: { type: ["string", "null"] }
        endereco: { type: ["string", "null"] }
        cidade: { type: ["string", "null"] }
        uf: { type: ["string", "null"] }
        ativo: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Aditivo:
      type: object
      required:
        - id
        - contratoId
        - tipo
        - numero
        - dataAssinatura
        - objeto
        - createdAt
      properties:
        id: { type: integer }
        contratoId: { type: integer }
        tipo:
          type: string
          enum: [aditivo, apostila, rescisao, suspensao]
        numero: { type: string }
        dataAssinatura: { type: string, format: date }
        dataNovaVigencia: { type: ["string", "null"], format: date }
        valorAcrescimo: { type: ["number", "null"] }
        objeto: { type: string }
        observacoes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    Medicao:
      type: object
      required:
        - id
        - contratoId
        - numero
        - mesReferencia
        - periodoInicio
        - periodoFim
        - valorMedido
        - valorPago
        - situacao
        - createdAt
      properties:
        id: { type: integer }
        contratoId: { type: integer }
        numero: { type: string }
        mesReferencia: { type: string, description: "MM/YYYY" }
        periodoInicio: { type: string, format: date }
        periodoFim: { type: string, format: date }
        valorMedido: { type: number }
        valorGlosa: { type: number }
        valorPago: { type: number }
        situacao:
          type: string
          enum: [pendente, em_analise, aprovada, paga, glosada, cancelada]
        numeroNotaFiscal: { type: ["string", "null"] }
        dataNotaFiscal: { type: ["string", "null"], format: date }
        observacoes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MedicaoInput:
      type: object
      required: [contratoId, numero, mesReferencia, periodoInicio, periodoFim, valorMedido]
      properties:
        contratoId: { type: integer }
        numero: { type: string }
        mesReferencia: { type: string, description: "MM/YYYY" }
        periodoInicio: { type: string, format: date }
        periodoFim: { type: string, format: date }
        valorMedido: { type: number }
        valorGlosa: { type: ["number", "null"] }
        numeroNotaFiscal: { type: ["string", "null"] }
        dataNotaFiscal: { type: ["string", "null"], format: date }
        observacoes: { type: ["string", "null"] }
    MedicaoUpdate:
      type: object
      properties:
        numero: { type: string }
        mesReferencia: { type: string }
        periodoInicio: { type: string, format: date }
        periodoFim: { type: string, format: date }
        valorMedido: { type: number }
        valorGlosa: { type: ["number", "null"] }
        situacao:
          type: string
          enum: [pendente, em_analise, aprovada, paga, glosada, cancelada]
        numeroNotaFiscal: { type: ["string", "null"] }
        dataNotaFiscal: { type: ["string", "null"], format: date }
        observacoes: { type: ["string", "null"] }
    MedicaoPagamentoInput:
      type: object
      required: [valorPago]
      properties:
        valorPago: { type: number }
        numeroNotaFiscal: { type: ["string", "null"] }
        dataNotaFiscal: { type: ["string", "null"], format: date }
        observacoes: { type: ["string", "null"] }

    # ── Alerta ────────────────────
    Alerta:
      type: object
      required: [id, contratoId, tipo, mensagem, lido, createdAt]
      properties:
        id: { type: integer }
        contratoId: { type: integer }
        contratoNumero: { type: ["string", "null"] }
        tipo:
          type: string
          enum: [vencimento_proximo, vencido, renovacao_necessaria, pendencia]
        mensagem: { type: string }
        lido: { type: boolean }
        createdAt: { type: string, format: date-time }

    # ── Dashboard ─────────────────
...
    PncpResultado:
      type: object
      required: [data, totalRegistros, totalPaginas, numeroPagina]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PncpContrato"
        totalRegistros: { type: integer }
        totalPaginas: { type: integer }
        numeroPagina: { type: integer }
    PncpContrato:
      type: object
      required:
        - numeroControlePNCP
        - objeto
        - orgaoEntidade
        - fornecedorNome
        - valorGlobal
        - dataAssinatura
        - dataVigenciaInicio
        - dataVigenciaFim
      properties:
        numeroControlePNCP: { type: string }
        numeroControlePncpCompra: { type: ["string", "null"] }
        anoContrato: { type: integer }
        numeroContratoEmpenho: { type: ["string", "null"] }
        processo: { type: ["string", "null"] }
        objeto: { type: string }
        orgaoEntidade:
          type: object
          properties:
            cnpj: { type: string }
            razaoSocial: { type: string }
        unidadeOrgao:
          type: object
          properties:
            codigoUnidade: { type: string }
            nomeUnidade: { type: string }
            ufSigla: { type: string }
            municipioNome: { type: string }
        fornecedorNome: { type: string }
        niFornecedor: { type: string }
        tipoPessoa: { type: string }
        tipoContrato:
          type: object
          properties:
            id: { type: integer }
            nome: { type: string }
        categoriaProcesso:
          type: object
          properties:
            id: { type: integer }
            nome: { type: string }
        valorInicial: { type: number }
        valorGlobal: { type: number }
        dataAssinatura: { type: string }
        dataVigenciaInicio: { type: string }
        dataVigenciaFim: { type: string }
        dataPublicacaoPncp: { type: string }
    PncpImportInput:
      type: object
      required: [numeroControlePNCP]
      properties:
        numeroControlePNCP: { type: string }
        fornecedorId: { type: ["integer", "null"] }
    OrgaoPncp:
      type: object
      required: [cnpj, razaoSocial]
      properties:
        cnpj: { type: string }
        razaoSocial: { type: string }
        nomeFantasia: { type: ["string", "null"] }
        poderId: { type: ["string", "null"] }
        esferaId: { type: ["string", "null"] }
        statusAtivo: { type: boolean }
...