# lib/api-zod/src/generated/types/pncpImportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SysCont Web - Agência Espacial Brasileira - Sistema de Gestão de Contratos Administrativos
 * OpenAPI spec version: 0.2.0
 */

export interface PncpImportInput {
  numeroControlePNCP: string;
  /** @nullable */
  fornecedorId?: number | null;
}